requirements.txt
openai>=1.0.0
pinecone>=3.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.3
pytesseract>=0.3.10
pandas>=2.0.0
tqdm>=4.66.0
python-multipart>=0.0.6
fastapi>=0.104.0
uvicorn>=0.24.0
python-magic>=0.4.27
Pillow>=10.0.0
pymupdf>=1.26.0
google-generativeai>=0.3.0
